+++ tmtccmd/cfdp/handler/source.py
@@ -46,7 +46,6 @@ from tmtccmd.cfdp import (
     RemoteEntityCfg,
 )
 from tmtccmd.cfdp.defs import CfdpState
-from tmtccmd.cfdp.filestore import VirtualFilestore
 from tmtccmd.cfdp.handler.crc import Crc32Helper
 from tmtccmd.cfdp.handler.defs import (
     FileParamsBase,
@@ -108,15 +107,40 @@ class _SourceFileParams(FileParamsBase):
 class SourceStateWrapper:
     state: CfdpState = CfdpState.IDLE
     step: TransactionStep = TransactionStep.IDLE
-    packets_ready: bool = False
+    _num_packets_ready: int = 0
+
+    @property
+    def num_packets_ready(self) -> int:
+        return self._num_packets_ready
+
+    @property
+    def packets_ready(self) -> bool:
+        return self.num_packets_ready > 0
+
+
+class _AckedModeParams:
+    def __init__(self) -> None:
+        self.step_before_retransmission: Optional[TransactionStep] = None
+        self.segment_reqs_to_handle: Optional[Tuple[int, int]] = None
+        self.segment_req_index: int = 0
+
+
+@dataclass
+class _PositiveAckProcedureParams:
+    ack_timer: Optional[Countdown] = None
+    ack_counter: int = 0
+    cond_code_of_eof_pdu: Optional[ConditionCode] = None
 
 
 class _TransferFieldWrapper:
-    def __init__(self, local_entity_id: UnsignedByteField, vfs: VirtualFilestore):
-        self.crc_helper = Crc32Helper(ChecksumType.NULL_CHECKSUM, vfs)
-        self.transaction: Optional[TransactionId] = None
+    def __init__(self, local_entity_id: UnsignedByteField):
+        self.transaction_id: Optional[TransactionId] = None
         self.check_timer: Optional[Countdown] = None
-        self.fp = _SourceFileParams.empty()
+        self.positive_ack_params: _PositiveAckProcedureParams = (
+            _PositiveAckProcedureParams()
+        )
+        self.ack_params: _AckedModeParams = _AckedModeParams()
+        self.fp: _SourceFileParams = _SourceFileParams.empty()
         self.finished_params: Optional[FinishedParams] = None
         self.remote_cfg: Optional[RemoteEntityCfg] = None
         self.closure_requested: bool = False
@@ -124,12 +148,16 @@ class _TransferFieldWrapper:
         self.pdu_conf.source_entity_id = local_entity_id
 
     @property
-    def source_id(self):
+    def source_id(self) -> UnsignedByteField:
         return self.pdu_conf.source_entity_id
 
     @source_id.setter
     def source_id(self, source_id: UnsignedByteField):
         self.pdu_conf.source_entity_id = source_id
+
+    @property
+    def positve_ack_counter(self) -> int:
+        return self.positive_ack_params.ack_counter
 
     @property
     def dest_id(self):
@@ -158,11 +186,12 @@ class _TransferFieldWrapper:
     def reset(self):
         self.fp.reset()
         self.remote_cfg = None
-        self.transaction = None
+        self.transaction_id = None
         self.check_timer = None
         self.closure_requested = False
         self.pdu_conf = PduConfig.empty()
         self.finished_params = None
+        self.positive_ack_params = _PositiveAckProcedureParams()
 
 
 class FsmResult:
@@ -194,21 +223,17 @@ class PduIgnoredAtSource(Exception):
         super().__init__(f"ignored PDU packet at source handler: {reason!r}")
 
 
-class _AckedModeContext:
-    def __init__(self) -> None:
-        self.segment_reqs_to_handle: Optional[Tuple[int, int]] = None
-        self.segment_req_index: int = 0
-
-
 class SourceHandler:
-    """This is the primary CFDP source handler. It models the CFDP source entity, which is primarily
-    responsible for handling put requests to send files to another CFDP destination entity.
+    """This is the primary CFDP source handler. It models the CFDP source entity, which is
+    primarily responsible for handling put requests to send files to another CFDP destination
+    entity.
 
     As such, it contains a state machine to perform all operations necessary to perform a
     source-to-destination file transfer. This class does not send the CFDP PDU packets directly
     to allow for greater flexibility. For example, a user might want to wrap the CFDP packet
     entities into a CCSDS space packet or into a special frame type. The user is responsible for
-    sending the packets and confirming that they are sent successfully.
+    sending the packets and confirming that they are sent successfully. The handler can handle
+    both unacknowledged (class 1) and acknowledged (class 2) file tranfers.
 
     The following core functions are the primary interface:
 
@@ -244,9 +269,9 @@ class SourceHandler:
         self.user = user
         self.seq_num_provider = seq_num_provider
         self.check_timer_provider = check_timer_provider
-        self._params = _TransferFieldWrapper(cfg.local_entity_id, self.user.vfs)
+        self._params = _TransferFieldWrapper(cfg.local_entity_id)
+        self._crc_helper = Crc32Helper(ChecksumType.NULL_CHECKSUM, self.user.vfs)
         self._put_req: Optional[PutRequest] = None
-        self._ack_ctx: _AckedModeContext = _AckedModeContext()
         self._inserted_pdu = PduHolder(None)
         self._pdus_to_be_sent: Deque[PduHolder] = deque()
 
@@ -268,6 +293,10 @@ class SourceHandler:
         self._params.source_id = source_id
 
     @property
+    def positive_ack_counter(self) -> int:
+        return self._params.positve_ack_counter
+
+    @property
     def state(self) -> CfdpState:
         return self.states.state
 
@@ -278,6 +307,10 @@ class SourceHandler:
     @property
     def packets_ready(self) -> bool:
         return self.states.packets_ready
+
+    @property
+    def num_packets_ready(self) -> int:
+        return self.states.num_packets_ready
 
     def put_request(self, request: PutRequest, remote_cfg: RemoteEntityCfg):
         """You can call this function to pass a put request to the source handler, which is
@@ -296,7 +329,7 @@ class SourceHandler:
         self._put_req = request
         self._params.remote_cfg = remote_cfg
         self._params.dest_id = remote_cfg.entity_id
-        self.states.packets_ready = False
+        self.states._num_packets_ready = 0
         self._setup_transmission_mode()
         if self._params.transmission_mode == TransmissionMode.UNACKNOWLEDGED:
             _LOGGER.debug("Starting Put Request handling in NAK mode")
@@ -328,8 +361,8 @@ class SourceHandler:
         if self.states.packets_ready:
             raise UnretrievedPdusToBeSent()
         if (
-            self._params.transaction is not None
-            and transaction_id == self._params.transaction
+            self._params.transaction_id is not None
+            and transaction_id == self._params.transaction_id
         ):
             self._declare_fault(ConditionCode.CANCEL_REQUEST_RECEIVED)
             return True
@@ -421,10 +454,9 @@ class SourceHandler:
 
     def get_next_packet(self) -> Optional[PduHolder]:
         """Retrieve the next packet which should be sent to the remote CFDP destination entity."""
-        if len(self._pdus_to_be_sent) <= 1:
-            self.states.packets_ready = False
         if len(self._pdus_to_be_sent) == 0:
             return None
+        self.states._num_packets_ready -= 1
         return self._pdus_to_be_sent.popleft()
 
     def state_machine(self) -> FsmResult:
@@ -454,7 +486,7 @@ class SourceHandler:
 
     @property
     def transaction_id(self) -> Optional[TransactionId]:
-        return self._params.transaction
+        return self._params.transaction_id
 
     def reset(self):
         """This function is public to allow completely resetting the handler, but it is explicitely
@@ -499,11 +531,11 @@ class SourceHandler:
         self._prepare_pdu_conf(file_size)
         self._get_next_transfer_seq_num()
         self._calculate_max_file_seg_len()
-        self._params.transaction = TransactionId(
+        self._params.transaction_id = TransactionId(
             source_entity_id=self.cfg.local_entity_id,
             transaction_seq_num=self.transaction_seq_num,
         )
-        self.user.transaction_indication(self._params.transaction)
+        self.user.transaction_indication(self._params.transaction_id)
 
     def _prepare_file_params(self):
         assert self._put_req is not None
@@ -575,7 +607,7 @@ class SourceHandler:
         if self._put_req.metadata_only:
             params = MetadataParams(
                 closure_requested=self._params.closure_requested,
-                checksum_type=self._params.crc_helper.checksum_type,
+                checksum_type=self._crc_helper.checksum_type,
                 file_size=0,
                 dest_file_name=None,
                 source_file_name=None,
@@ -602,7 +634,7 @@ class SourceHandler:
         return MetadataParams(
             dest_file_name=self._put_req.dest_file.as_posix(),  # type: ignore
             source_file_name=self._put_req.source_file.as_posix(),  # type: ignore
-            checksum_type=self._params.crc_helper.checksum_type,
+            checksum_type=self._crc_helper.checksum_type,
             closure_requested=self._params.closure_requested,
             file_size=self._params.fp.file_size,
         )
@@ -615,7 +647,6 @@ class SourceHandler:
             if self.__handle_retransmission():
                 return True
         if self._prepare_progressing_file_data_pdu():
-            self.states.packets_ready = True
             return True
         if self._params.fp.no_eof:
             # Special case: Metadata Only.
@@ -653,13 +684,15 @@ class SourceHandler:
                 )
                 packet_prepared = True
         if packet_prepared:
+            self._params.ack_params.step_before_retransmission = self.states.step
             self.states.step = TransactionStep.RETRANSMITTING
+        self._inserted_pdu.pdu = None
         return packet_prepared
 
     def _handle_waiting_for_ack(self):
         if self.states.state != CfdpState.BUSY_CLASS_2_ACKED:
             _LOGGER.error(
-                f"Invalid ACK waiting function call for state {self.states.state}"
+                f"invalid ACK waiting function call for state {self.states.state}"
             )
 
         if self.__handle_retransmission():
@@ -668,17 +701,39 @@ class SourceHandler:
             self._inserted_pdu.pdu_type == PduType.FILE_DIRECTIVE
             and self._inserted_pdu.pdu_directive_type != DirectiveType.ACK_PDU
         ):
+            self._handle_positive_ack_procedures()
             return
-        # TODO: Implement Positive Acknowledgement Procedures according to 4.7.1
         ack_pdu = self._inserted_pdu.to_ack_pdu()
         if ack_pdu.directive_code_of_acked_pdu == DirectiveType.EOF_PDU:
-            if ack_pdu.condition_code_of_acked_pdu != ConditionCode.NO_ERROR:
-                # TODO: This is required for class 2 transfers. It might make sense
-                #       to remember the condition code of the sent EOF PDU for a basic
-                #       equality check here
-                pass
+            # TODO: Equality check required? I am not sure why the condition code is supplied
+            #       as part of the ACK packet.
             self.states.step = TransactionStep.WAITING_FOR_FINISHED
+        else:
+            _LOGGER.error(
+                f"received ACK PDU with invalid acked directive code"
+                f" {ack_pdu.directive_code_of_acked_pdu!r}"
+            )
+
         self._inserted_pdu.pdu = None
+
+    def _handle_positive_ack_procedures(self):
+        """Positive ACK procedures according to chapter 4.7.1 of the CFDP standard."""
+        assert self._params.positive_ack_params.ack_timer is not None
+        assert self._params.remote_cfg is not None
+        if self._params.positive_ack_params.ack_timer.timed_out():
+            assert self._params.positive_ack_params.cond_code_of_eof_pdu is not None
+            if (
+                self._params.positive_ack_params.ack_counter
+                == self._params.remote_cfg.positive_ack_timer_expiration_limit
+            ):
+                self._declare_fault(ConditionCode.POSITIVE_ACK_LIMIT_REACHED)
+                return
+            self._params.positive_ack_params.ack_timer.reset()
+            self._params.positive_ack_params.ack_counter += 1
+            self._prepare_eof_pdu(
+                self._params.positive_ack_params.cond_code_of_eof_pdu,
+                self._checksum_calculation(self._params.fp.file_size),
+            )
 
     def _handle_wait_for_finish(self):
         if (
@@ -714,7 +769,7 @@ class SourceHandler:
 
     def _notice_of_completion(self):
         if self.cfg.indication_cfg.transaction_finished_indication_required:
-            assert self._params.transaction is not None
+            assert self._params.transaction_id is not None
             # This happens for unacknowledged file copy operation with no closure.
             if self._params.finished_params is None:
                 self._params.finished_params = FinishedParams(
@@ -723,7 +778,7 @@ class SourceHandler:
                     ConditionCode.NO_ERROR,
                 )
             indication_params = TransactionFinishedParams(
-                transaction_id=self._params.transaction,
+                transaction_id=self._params.transaction_id,
                 finished_params=self._params.finished_params,
             )
             self.user.transaction_finished_indication(indication_params)
@@ -739,7 +794,8 @@ class SourceHandler:
         if self.states.step == TransactionStep.SENDING_METADATA:
             self.states.step = TransactionStep.SENDING_FILE_DATA
         elif self.states.step == TransactionStep.RETRANSMITTING:
-            self.states.step = TransactionStep.SENDING_FILE_DATA
+            assert self._params.ack_params.step_before_retransmission is not None
+            self.states.step = self._params.ack_params.step_before_retransmission
         elif self.states.step == TransactionStep.SENDING_FILE_DATA:
             self._handle_file_data_sent()
         elif self.states.step == TransactionStep.SENDING_ACK_OF_FINISHED:
@@ -749,8 +805,8 @@ class SourceHandler:
 
     def _handle_eof_sent(self):
         if self.cfg.indication_cfg.eof_sent_indication_required:
-            assert self._params.transaction is not None
-            self.user.eof_sent_indication(self._params.transaction)
+            assert self._params.transaction_id is not None
+            self.user.eof_sent_indication(self._params.transaction_id)
         if self.states.state == CfdpState.BUSY_CLASS_1_NACKED:
             if self._params.closure_requested:
                 assert self._params.remote_cfg is not None
@@ -765,7 +821,7 @@ class SourceHandler:
             else:
                 self.states.step = TransactionStep.NOTICE_OF_COMPLETION
         else:
-            self.states.step = TransactionStep.WAITING_FOR_EOF_ACK
+            self._start_positive_ack_procedure()
 
     def _handle_file_data_sent(self):
         if self._params.fp.progress == self._params.fp.file_size:
@@ -779,6 +835,14 @@ class SourceHandler:
             TransactionStatus.ACTIVE,
         )
         self._add_packet_to_be_sent(ack_pdu)
+
+    def _start_positive_ack_procedure(self):
+        assert self._params.remote_cfg is not None
+        self.states.step = TransactionStep.WAITING_FOR_EOF_ACK
+        self._params.positive_ack_params.ack_timer = Countdown.from_seconds(
+            self._params.remote_cfg.positive_ack_timer_interval_seconds
+        )
+        self._params.positive_ack_params.ack_counter = 0
 
     def _setup_transmission_mode(self):
         assert self._put_req is not None
@@ -794,11 +858,11 @@ class SourceHandler:
         self._params.transmission_mode = trans_mode_to_set
         # This also sets the field of the PDU configuration struct.
         self._params.closure_requested = closure_req_to_set
-        self._params.crc_helper.checksum_type = self._params.remote_cfg.crc_type
+        self._crc_helper.checksum_type = self._params.remote_cfg.crc_type
 
     def _add_packet_to_be_sent(self, packet: GenericPduPacket):
         self._pdus_to_be_sent.append(PduHolder(packet))
-        self.states.packets_ready = True
+        self.states._num_packets_ready += 1
 
     def _prepare_progressing_file_data_pdu(self) -> bool:
         """Prepare the next file data PDU, which also progresses the file copy operation.
@@ -846,6 +910,7 @@ class SourceHandler:
             self._add_packet_to_be_sent(file_data_pdu)
 
     def _prepare_eof_pdu(self, condition_code: ConditionCode, checksum: bytes):
+        self._params.positive_ack_params.cond_code_of_eof_pdu = condition_code
         self._add_packet_to_be_sent(
             EofPdu(
                 file_checksum=checksum,
@@ -877,6 +942,19 @@ class SourceHandler:
         self.cfg.default_fault_handlers.report_fault(cond)
 
     def _notice_of_cancellation(self, condition_code: ConditionCode):
+        # CFDP standard 4.11.2.2.3: Any fault declared in the course of transferring
+        # the EOF (cancel) PDU must result in abandonment of the transaction.
+        if (
+            self._params.positive_ack_params.cond_code_of_eof_pdu is not None
+            and self._params.positive_ack_params.cond_code_of_eof_pdu
+            != ConditionCode.NO_ERROR
+        ):
+            # We still call the abandonment callback to ensure the fault is logged.
+            self.cfg.default_fault_handlers.abandoned_cb(
+                self._params.positive_ack_params.cond_code_of_eof_pdu
+            )
+            self._abandon_transaction()
+            return
         # As specified in 4.11.2.2, prepare an EOF PDU to be sent to the remote entity. Supply
         # the checksum for the file copy progress sent so far.
         self._prepare_eof_pdu(
@@ -900,7 +978,7 @@ class SourceHandler:
         else:
             assert self._put_req is not None
             assert self._put_req.source_file is not None
-            crc = self._params.crc_helper.calc_for_file(
+            crc = self._crc_helper.calc_for_file(
                 file=self._put_req.source_file,
                 file_sz=size_to_calculate,
                 segment_len=self._params.fp.segment_len,