+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Starting from v4.0.0, this project adheres to [Semantic Versioning](http://semve
 ## Added
 
 - The CFDP source and destination handler can now also handle cancel requests.
+- The CFDP source handler can handle acknowleded mode transfers now.
+- New `from_seconds` classmethod for `Countdown` class.
 
 ## Changed